Add _blank targets on reporting page
When on reporting page, reporter might be in the middle of writing a
long(er) reporting information and clicks on one of the given links.
Instead of opening the page in the same window, click should open a new
tab or window instead to possibly not loosing entered bug report info.