$pseudo_pkgs now has an array of children as a third element.
If parent = '0' in the database, $pseudo_pkgs[2] is an array. If it is
nonzero (refers to a valid id of another row), $pseudo_pkgs[2] is a
child and is null. We no longer need &nbsp;&nbsp;&nbsp;&nbsp; to
distinguish between parents and children. Nowhere do we count the length
of $pseudo_pkgs, so the extra element should not break anything.