Fix 76079: Attaching PRs to bug issues broken
This patch fixes attaching pull requests to bug tickets. This will need
to be resolved also using different consuming of the GitHub API in the
near future due to possible rate limits and other limitations.
Additionally, would be good to get the PEAR dependency refactored for
handling the errors to not depend on outdated PEAR.

Otherwise, this fixes bug #76079 thanks to @zrhoffman