Fix #76961: vote throws a Server Error over ipv6
This patch fixes a bug when user with IPv6 is voting on a bug. It syncs
the functionality with the rest of the ip column types in the database.

Before applying this patch, the database schema need to be updated and
current IPs should ideally be converted from integers to numeric values
returned by the INET6_ATON() MySQL function.