fix #81460 Bad validation of input parameters of report.php
a POST parameter was sent to a hash function that only accepts string as parameter. By manipulation of that POST parameter was possible to pass an array which resulted in full path disclosure in a default install with PHP8+ without silencing error_reporting etc.

Other places for the validation code also possible, this is the quick solution I come up with.